requirements.txt
aiogram==2.25.1
asyncpg==0.29.0
python-dateutil==2.9.0